refactor: strategy 전체 제거 및 piece에 이동로직 적용
