+++ src/main/java/chess/domain/piece/pawn/Pawn.java
@@ -1,21 +1,60 @@
 package chess.domain.piece.pawn;
 
+import chess.domain.ChessBoard;
 import chess.domain.Color;
+import chess.domain.Position;
+import chess.domain.direction.Direction;
 import chess.domain.piece.Piece;
-import chess.domain.piece.strategy.PawnMovableStrategy;
-import chess.domain.piece.strategy.WhitePawnFirstMovableStrategy;
+import java.util.List;
 
 public abstract class Pawn extends Piece {
 
     private static final String PAWN_NAME = "P";
     private static final double PAWN_SCORE = 1;
+    private static final int MOVABLE_TO_ENEMY_COUNT = 1;
 
-    public Pawn(Color color) {
-        super(color, PAWN_NAME, new WhitePawnFirstMovableStrategy());
+    private final int movableCount;
+    private final Direction moveDirection;
+    private final List<Direction> moveDirectionToEnemy;
+
+    protected Pawn(Color color, int movableCount, Direction moveDirection, List<Direction> moveDirectionToEnemy) {
+        super(color, PAWN_NAME);
+        this.movableCount = movableCount;
+        this.moveDirection = moveDirection;
+        this.moveDirectionToEnemy = moveDirectionToEnemy;
     }
 
-    public Pawn(Color color, PawnMovableStrategy pawnMovableStrategy) {
-        super(color, PAWN_NAME, pawnMovableStrategy);
+    @Override
+    public final boolean isMovable(Position source, Position target, ChessBoard chessBoard) {
+        if (existDirectionToEnemyPiecePosition(source, target, chessBoard)) {
+            return isMovableToEnemyPosition(source, target);
+        }
+        return isMovableToPosition(source, target, chessBoard);
+    }
+
+    private boolean existDirectionToEnemyPiecePosition(Position source, Position target, ChessBoard chessBoard) {
+        Piece piece = chessBoard.pieceByPosition(source);
+        return !chessBoard.isPositionEmpty(target) && !piece.isSameTeamPiece(chessBoard.pieceByPosition(target));
+    }
+
+    private boolean isMovableToEnemyPosition(Position source, Position target) {
+        return moveDirectionToEnemy.stream()
+                .map(direction -> direction.route(source, target))
+                .anyMatch(positions -> positions.size() == MOVABLE_TO_ENEMY_COUNT);
+    }
+
+    private boolean isMovableToPosition(Position source, Position target, ChessBoard chessBoard) {
+        List<Position> route = moveDirection.route(source, target);
+        return existMovablePosition(route) && isClearRoute(route, chessBoard);
+    }
+
+    private boolean existMovablePosition(List<Position> route) {
+        return !route.isEmpty() && route.size() <= movableCount;
+    }
+
+    private boolean isClearRoute(List<Position> route, ChessBoard chessBoard) {
+        return route.stream()
+                .allMatch(chessBoard::isPositionEmpty);
     }
 
     @Override
