--- src/main/java/chess/domain/piece/strategy/PawnMovableStrategy.java
@@ -1,55 +1,0 @@
-package chess.domain.piece.strategy;
-
-import chess.domain.ChessBoard;
-import chess.domain.Position;
-import chess.domain.direction.Direction;
-import chess.domain.piece.Piece;
-import java.util.List;
-
-public abstract class PawnMovableStrategy implements PieceMovableStrategy {
-
-    private static final int MOVABLE_TO_ENEMY_COUNT = 1;
-
-    private final int movableCount;
-    private final Direction moveDirection;
-    private final List<Direction> moveDirectionToEnemy;
-
-    protected PawnMovableStrategy(int movableCount, Direction moveDirection, List<Direction> moveDirectionToEnemy) {
-        this.movableCount = movableCount;
-        this.moveDirection = moveDirection;
-        this.moveDirectionToEnemy = moveDirectionToEnemy;
-    }
-
-    @Override
-    public boolean isMovable(Position source, Position target, ChessBoard chessBoard) {
-        if (existDirectionToEnemyPiecePosition(source, target, chessBoard)) {
-            return isMovableToEnemyPosition(source, target);
-        }
-        return isMovableToPosition(source, target, chessBoard);
-    }
-
-    private boolean existDirectionToEnemyPiecePosition(Position source, Position target, ChessBoard chessBoard) {
-        Piece piece = chessBoard.pieceByPosition(source);
-        return !chessBoard.isPositionEmpty(target) && !piece.isSameTeamPiece(chessBoard.pieceByPosition(target));
-    }
-
-    private boolean isMovableToEnemyPosition(Position source, Position target) {
-        return moveDirectionToEnemy.stream()
-                .map(direction -> direction.route(source, target))
-                .anyMatch(positions -> positions.size() == MOVABLE_TO_ENEMY_COUNT);
-    }
-
-    private boolean isMovableToPosition(Position source, Position target, ChessBoard chessBoard) {
-        List<Position> route = moveDirection.route(source, target);
-        return existMovablePosition(route) && isClearRoute(route, chessBoard);
-    }
-
-    private boolean existMovablePosition(List<Position> route) {
-        return !route.isEmpty() && route.size() <= movableCount;
-    }
-
-    private boolean isClearRoute(List<Position> route, ChessBoard chessBoard) {
-        return route.stream()
-                .allMatch(chessBoard::isPositionEmpty);
-    }
-}
