--- src/main/java/chess/domain/piece/strategy/PieceMovableStrategy.java
@@ -1,9 +1,0 @@
-package chess.domain.piece.strategy;
-
-import chess.domain.ChessBoard;
-import chess.domain.Position;
-
-public interface PieceMovableStrategy {
-
-    boolean isMovable(Position source, Position target, ChessBoard chessBoard);
-}
