+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,13 +1,28 @@
 package chess.domain.piece;
 
+import static chess.domain.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.ChessBoard;
 import chess.domain.Color;
+import chess.domain.Position;
+import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.EnumSource;
 
 class KingTest {
+
+    private Piece king;
+    private Position source;
+
+    @BeforeEach
+    void setUp() {
+        king = new King(Color.WHITE);
+        source = new Position('a', '1');
+    }
 
     @ParameterizedTest
     @EnumSource(Color.class)
@@ -15,5 +30,28 @@ class KingTest {
     void createKing(Color color) {
         Piece king = new King(color);
         assertThat(king).isInstanceOf(King.class);
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {"a,2,true", "b,1,true", "b,2,true", "b,3,false", "c,3,false"})
+    @DisplayName("킹의 빈곳 이동 가능 여부 확인")
+    void isMovableToEmptyPosition(char col, char row, boolean expected) {
+        Position target = new Position(col, row);
+        ChessBoard chessBoard = new ChessBoard(Map.of(source, new King(WHITE)));
+
+        assertThat(king.isMovable(source, target, chessBoard)).isEqualTo(expected);
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {"WHITE,false", "BLACK,true"})
+    @DisplayName("기물이 존재할 경우의 이동 가능 여부 확인")
+    void isMovableToDifferentPiecePosition(Color color, boolean expected) {
+        Position target = new Position('a', '2');
+        ChessBoard chessBoard = new ChessBoard(Map.of(
+                source, new King(WHITE),
+                target, new King(color)
+        ));
+
+        assertThat(king.isMovable(source, target, chessBoard)).isEqualTo(expected);
     }
 }
