+++ src/test/java/chess/domain/piece/pawn/BlackPawnTest.java
@@ -1,12 +1,12 @@
-package chess.domain.piece.strategy;
+package chess.domain.piece.pawn;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
 import chess.domain.ChessBoard;
 import chess.domain.Position;
-import chess.domain.piece.pawn.BlackFirstPawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.pawn.WhiteFirstPawn;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
@@ -16,25 +16,27 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class WhitePawnFirstMovableStrategyTest {
+class BlackPawnTest {
 
-    private PieceMovableStrategy whitePawnFirstMovableStrategy;
+    private Piece pawn;
     private Position source;
 
     @BeforeEach
     void setUp() {
-        whitePawnFirstMovableStrategy = new WhitePawnFirstMovableStrategy();
-        source = new Position('b', '1');
+        source = new Position('b', '7');
+        pawn = new BlackPawn().move(new Position('b', '8'), source, new ChessBoard(Map.of(
+                new Position('b', '8'), new BlackPawn()
+        )));
     }
 
     @ParameterizedTest
-    @CsvSource(value = {"b,2,true", "b,3,true", "b,1,false", "b,4,false", "a,1,false", "a,2,false"})
+    @CsvSource(value = {"b,6,true", "b,5,false", "b,7,false", "a,8,false", "a,6,false"})
     @DisplayName("폰의 빈곳 전진 가능 여부 확인")
     void isMovableToEmptyPosition(char col, char row, boolean expected) {
         Position target = new Position(col, row);
-        ChessBoard chessBoard = new ChessBoard(Map.of(source, new WhiteFirstPawn()));
+        ChessBoard chessBoard = new ChessBoard(Map.of(source, new WhitePawn()));
 
-        assertThat(whitePawnFirstMovableStrategy.isMovable(source, target, chessBoard)).isEqualTo(expected);
+        assertThat(pawn.isMovable(source, target, chessBoard)).isEqualTo(expected);
     }
 
     @ParameterizedTest
@@ -42,18 +44,16 @@ class WhitePawnFirstMovableStrategyTest {
     @DisplayName("기물이 가로막을 경우의 전진 불가능")
     void cannotMoveToPiecePosition(Position target, Piece piece) {
         ChessBoard chessBoard = new ChessBoard(Map.of(
-                source, new WhiteFirstPawn(),
+                source, new WhitePawn(),
                 target, piece));
 
-        assertThat(whitePawnFirstMovableStrategy.isMovable(source, target, chessBoard)).isFalse();
+        assertThat(pawn.isMovable(source, target, chessBoard)).isFalse();
     }
 
     private static Stream<Arguments> cannotMoveToPiecePosition() {
         return Stream.of(
-                Arguments.of(new Position('b', '2'), new WhiteFirstPawn()),
-                Arguments.of(new Position('b', '3'), new WhiteFirstPawn()),
-                Arguments.of(new Position('b', '2'), new BlackFirstPawn()),
-                Arguments.of(new Position('b', '3'), new BlackFirstPawn())
+                Arguments.of(new Position('b', '6'), new WhitePawn()),
+                Arguments.of(new Position('b', '6'), new BlackPawn())
         );
     }
 
@@ -62,18 +62,18 @@ class WhitePawnFirstMovableStrategyTest {
     @DisplayName("대각선 방향에 적이 있으면 전진 가능")
     void canMoveToEnemyPiecePosition(Position target, Piece piece, boolean expected) {
         ChessBoard chessBoard = new ChessBoard(Map.of(
-                source, new WhiteFirstPawn(),
+                source, new WhitePawn(),
                 target, piece));
 
-        assertThat(whitePawnFirstMovableStrategy.isMovable(source, target, chessBoard)).isEqualTo(expected);
+        assertThat(pawn.isMovable(source, target, chessBoard)).isEqualTo(expected);
     }
 
     private static Stream<Arguments> canMoveToEnemyPiecePosition() {
         return Stream.of(
-                Arguments.of(new Position('a', '2'), new BlackFirstPawn(), true),
-                Arguments.of(new Position('c', '2'), new BlackFirstPawn(), true),
-                Arguments.of(new Position('a', '2'), new WhiteFirstPawn(), false),
-                Arguments.of(new Position('c', '2'), new WhiteFirstPawn(), false)
+                Arguments.of(new Position('a', '6'), new BlackPawn(), true),
+                Arguments.of(new Position('c', '6'), new BlackPawn(), true),
+                Arguments.of(new Position('a', '6'), new WhitePawn(), false),
+                Arguments.of(new Position('c', '6'), new WhitePawn(), false)
         );
     }
 }
