+++ src/test/java/chess/domain/piece/pawn/PawnTest.java
@@ -3,8 +3,6 @@ package chess.domain.piece.pawn;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.piece.Piece;
-import chess.domain.piece.pawn.Pawn;
-import chess.domain.piece.pawn.WhiteFirstPawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,7 +11,7 @@ class PawnTest {
     @Test
     @DisplayName("폰 기물 생성")
     void createPawn() {
-        Piece pawn = new WhiteFirstPawn();
+        Piece pawn = new WhitePawn();
         assertThat(pawn).isInstanceOf(Pawn.class);
     }
 }
